Move Links into context module
